Removed build warnings and fixed other issues
