--- src/py_assignment.h
@@ -1,17 +1,0 @@
-#ifndef PY_ASSIGNMENT_H
-#define PY_ASSIGNMENT_H
-
-#include <Python.h>
-#include <structmember.h>
-
-typedef struct {
-    PyObject_HEAD
-    int n;
-    unsigned short *assignment;
-} AssignmentObject;
-
-extern PyTypeObject AssignmentType;
-
-void init_assignment_type_fields(void);
-
-#endif
