--- src/py_fastkmeans_methods.h
@@ -1,4 +1,0 @@
-#include <Python.h>
-#include <structmember.h>
-
-extern PyMethodDef Fastkmeans_methods[];
