+++ src/python-bindings/fastkmeans.cpp
@@ -1,5 +1,12 @@
+/* Creates the fastkmeans Python module and adds the classes and module methods.
+ * This module provides wrappers for the Kmeans algorithms in the fast-kmeans
+ * library, as well as auxiliary classes.
+ */
+
 #include <Python.h> // Provides Python/C API 
 #include <structmember.h> // Provides utilities for dealing with attributes
+
+#include <cstring> // For strrchr
 
 #include "py_annulus.h"
 #include "py_assignment.h"
@@ -12,8 +19,6 @@
 #include "py_fastkmeans_methods.h"
 #include "py_naive.h"
 #include "py_sort.h"
-
-#include <cstring>
 
 extern "C" {
     static PyTypeObject *type_object_ptrs[] = {
@@ -31,42 +36,32 @@ extern "C" {
     };
 
     static PyModuleDef fastkmeansmodule = {
-        PyModuleDef_HEAD_INIT, // .m_base: required
-        "fastkmeans", // .m_name
-        "Extension module for the fast-kmeans C++ library", // .m_doc
-        -1, // .m_size
-        Fastkmeans_methods, // .m_methods
-        // .m_slots
-        // .m_traverse
-        // .m_clear
-        // .m_free
+        PyModuleDef_HEAD_INIT, // m_base: required
+        "fastkmeans", // m_name
+        "Extension module for the fast-kmeans C++ library", // m_doc
+        -1, // m_size
+        Fastkmeans_methods, // m_methods
+        NULL, // m_slots
+        NULL, // m_traverse
+        NULL, // m_clear
+        NULL, // m_free
     };
 
     PyMODINIT_FUNC PyInit_fastkmeans(void) {
-        // TODO make each of these return PyTypeObject * instead?
-
-        init_annulus_type_fields();
-        init_assignment_type_fields();
-        init_compare_type_fields();
-        init_dataset_type_fields();
-        init_drake_type_fields();
-        init_elkan_type_fields();
-        init_hamerly_type_fields();
-        init_heap_type_fields();
-        init_naive_type_fields();
-        init_sort_type_fields();
-
+        // Initialize wrapper type structs
         for (int i = 0; type_object_ptrs[i] != NULL; i++) {
             if (PyType_Ready(type_object_ptrs[i]) < 0) {
                 return NULL;
             }
         }
 
+        // Initialize fastkmeans module
         PyObject *mod = PyModule_Create(&fastkmeansmodule);
         if (mod == NULL) {
             return NULL;
         }
 
+        // Add each wrapper type object to the module
         for (int i = 0; type_object_ptrs[i] != NULL; i++) {
             // Extract class name from fully-qualified name
             const char *name = strrchr(type_object_ptrs[i]->tp_name, '.') + 1;
