+++ src/python-bindings/py_drake.cpp
@@ -1,17 +1,20 @@
-#include "py_elkan.h"
-#include "py_dataset.h"
+/* DrakeKmeans wrapper. The comment at the beginning of each function definition
+ * demonstrates its usage in Python.
+ */
+
+#include "py_drake.h"
+
 #include "py_assignment.h"
+#include "py_dataset.h"
 
-//#include "elkan_kmeans.h"
 
-
-// Elkan instance object
+// Drake instance object
 
 
 /*
 typedef struct {
     PyObject_HEAD
-    ElkanKmeans *instance;
+    DrakeKmeans *instance;
 
     PyObject *dataset;
     PyObject *assignment;
@@ -20,18 +23,22 @@ typedef struct {
     // #ifdef COUNT_DISTANCES
     // long long num_distances;
     // #endif
-} ElkanObject;
+} DrakeObject;
 */
 
 
 // Object special methods
 
 
-static int Elkan_init(ElkanObject *self) {
-    // Elkan()
+static int Drake_init(DrakeObject *self, PyObject *args) {
+    // Drake()
 
-    self->instance = new ElkanKmeans();
+    int numBounds;
+    if (!PyArg_ParseTuple(args, "i", &numBounds)) {
+        return -1;
+    }
 
+    self->instance = new DrakeKmeans(numBounds);
     Py_INCREF(Py_None);
     self->dataset = Py_None;
     Py_INCREF(Py_None);
@@ -40,9 +47,8 @@ static int Elkan_init(ElkanObject *self) {
     return 0;
 }
 
-static void Elkan_dealloc(ElkanObject *self) {
+static void Drake_dealloc(DrakeObject *self) {
     delete self->instance;
-
     Py_DECREF(self->dataset);
     self->dataset = NULL;
     Py_DECREF(self->assignment);
@@ -55,8 +61,8 @@ static void Elkan_dealloc(ElkanObject *self) {
 // Object properties
 
 
-static PyObject * Elkan_get_centers(ElkanObject *self, void *closure) {
-    // a_elkan.centers
+static PyObject * Drake_get_centers(DrakeObject *self, void *closure) {
+    // a_drake.centers
 
     const Dataset *centers = self->instance->getCenters();
     if (centers == NULL) {
@@ -73,47 +79,57 @@ static PyObject * Elkan_get_centers(ElkanObject *self, void *closure) {
 
     // Copy values from centers to preserve constness
 
-    // Dataset *newCenters = new Dataset(centers->n, centers->d);
     for (int i = 0; i < centers->nd; i++) {
-        //newCenters->data[i] = centers->data[i];
         centersObj->dataset->data[i] = centers->data[i];
     }
-
-    //centersObj->dataset = newCenters;
 
     return (PyObject *) centersObj;
 }
 
-static PyGetSetDef Elkan_getsetters[] = {
-    // Readonly
-    {"centers", (getter) Elkan_get_centers, NULL, "The set of centers"},
+static PyGetSetDef Drake_getsetters[] = {
+    {
+        const_cast<char *>("centers"),
+        (getter) Drake_get_centers,
+        NULL, // Setter
+        const_cast<char *>("The set of centers"),
+    },
     {NULL} // Sentinel
 };
 
 
-// ElkanKmeans methods
+// DrakeKmeans methods
 
 
-static PyObject * Elkan_initialize(ElkanObject *self, PyObject *args,
+static PyObject * Drake_free(DrakeObject *self) {
+    // a_drake.free()
+
+    self->instance->free();
+    Py_RETURN_NONE;
+}
+
+static PyObject * Drake_initialize(DrakeObject *self, PyObject *args,
         PyObject *kwargs) {
-    // a_elkan.initialize(x, k, initial_assignment, num_threads=an_int)
+    // a_drake.initialize(x, k, initial_assignment, num_threads=an_int)
 
     PyObject *x_orig, *initAssigns_orig;
     unsigned short k;
     int numThreads = 1;
 
-    char *kwlist[] = {"", "", "", "num_threads", NULL};
-
+    char *emptyStr = const_cast<char *>("");
+    char *kwlist[] = {emptyStr, emptyStr, emptyStr,
+        const_cast<char *>("num_threads"), NULL};
     if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O!HO!|i", kwlist,
                 &DatasetType, &x_orig, &k, &AssignmentType, &initAssigns_orig,
                 &numThreads)) {
         return NULL;
     }
 
+    // Reassign dataset
     Py_DECREF(self->dataset);
     Py_INCREF(x_orig);
     self->dataset = x_orig;
 
+    // Reassign assignment
     Py_DECREF(self->assignment);
     Py_INCREF(initAssigns_orig);
     self->assignment = initAssigns_orig;
@@ -130,15 +146,8 @@ static PyObject * Elkan_initialize(ElkanObject *self, PyObject *args,
     Py_RETURN_NONE;
 }
 
-static PyObject * Elkan_free(ElkanObject *self) {
-    // a_elkan.free()
-
-    self->instance->free();
-    Py_RETURN_NONE;
-}
-
-static PyObject * Elkan_get_name(ElkanObject *self) {
-    // a_elkan.get_name()
+static PyObject * Drake_get_name(DrakeObject *self) {
+    // a_drake.get_name()
 
     return PyUnicode_FromString(self->instance->getName().c_str());
 }
@@ -147,12 +156,11 @@ static PyObject * Elkan_get_name(ElkanObject *self) {
 // OriginalSpaceKmeans methods
 
 
-static PyObject * Elkan_point_point_inner_product(ElkanObject *self,
+static PyObject * Drake_point_point_inner_product(DrakeObject *self,
         PyObject *args) {
-    // a_elkan.point_point_inner_product(x1, x2)
+    // a_drake.point_point_inner_product(x1, x2)
 
     int x1ndx, x2ndx;
-
     if (!PyArg_ParseTuple(args, "ii", &x1ndx, &x2ndx)) {
         return NULL;
     }
@@ -162,13 +170,12 @@ static PyObject * Elkan_point_point_inner_product(ElkanObject *self,
     return PyFloat_FromDouble(innerProd);
 }
 
-static PyObject * Elkan_point_center_inner_product(ElkanObject *self,
+static PyObject * Drake_point_center_inner_product(DrakeObject *self,
         PyObject *args) {
-    // a_elkan.point_center_inner_product(xndx, cndx)
+    // a_drake.point_center_inner_product(xndx, cndx)
 
     int xndx;
     unsigned short cndx;
-
     if (!PyArg_ParseTuple(args, "iH", &xndx, &cndx)) {
         return NULL;
     }
@@ -178,12 +185,11 @@ static PyObject * Elkan_point_center_inner_product(ElkanObject *self,
     return PyFloat_FromDouble(innerProd);
 }
 
-static PyObject * Elkan_center_center_inner_product(ElkanObject *self,
+static PyObject * Drake_center_center_inner_product(DrakeObject *self,
         PyObject *args) {
-    // a_elkan.center_center_inner_product(c1, c2)
+    // a_drake.center_center_inner_product(c1, c2)
 
     unsigned short c1, c2;
-
     if (!PyArg_ParseTuple(args, "HH", &c1, &c2)) {
         return NULL;
     }
@@ -197,13 +203,13 @@ static PyObject * Elkan_center_center_inner_product(ElkanObject *self,
 // Kmeans methods
 
 
-static PyObject * Elkan_run(ElkanObject *self, PyObject *args,
+static PyObject * Drake_run(DrakeObject *self, PyObject *args,
         PyObject *kwargs) {
-    // a_elkan.run(max_iterations = 0)
+    // a_drake.run(max_iterations = 0)
 
-    static char *kwlist[] = {"max_iterations"};
     int maxIterations = 0;
 
+    static char *kwlist[] = {const_cast<char *>("max_iterations"), NULL};
     if (!PyArg_ParseTupleAndKeywords(args, kwargs, "|i", kwlist,
                 &maxIterations)) {
         return NULL;
@@ -215,11 +221,10 @@ static PyObject * Elkan_run(ElkanObject *self, PyObject *args,
     return PyLong_FromLong(numIterations);
 }
 
-static PyObject * Elkan_get_assignment(ElkanObject *self, PyObject *args) {
-    // a_elkan.get_assignment(xndx)
+static PyObject * Drake_get_assignment(DrakeObject *self, PyObject *args) {
+    // a_drake.get_assignment(xndx)
 
     int xndx;
-
     if (!PyArg_ParseTuple(args, "i", &xndx)) {
         return NULL;
     }
@@ -229,11 +234,10 @@ static PyObject * Elkan_get_assignment(ElkanObject *self, PyObject *args) {
     return PyLong_FromLong(assignment);
 }
 
-static PyObject * Elkan_verify_assignment(ElkanObject *self, PyObject *args) {
-    // a_elkan.verify_assignment(iteration, startndx, endndx)
+static PyObject * Drake_verify_assignment(DrakeObject *self, PyObject *args) {
+    // a_drake.verify_assignment(iteration, startndx, endndx)
 
     int iteration, startNdx, endNdx;
-
     if (!PyArg_ParseTuple(args, "iii", &iteration, &startNdx, &endNdx)) {
         return NULL;
     }
@@ -241,18 +245,17 @@ static PyObject * Elkan_verify_assignment(ElkanObject *self, PyObject *args) {
     Py_RETURN_NONE;
 }
 
-static PyObject * Elkan_get_sse(ElkanObject *self) {
-    // a_elkan.get_sse()
+static PyObject * Drake_get_sse(DrakeObject *self) {
+    // a_drake.get_sse()
 
     return PyFloat_FromDouble(self->instance->getSSE());
 }
 
-static PyObject * Elkan_point_center_dist_2(ElkanObject *self, PyObject *args) {
-    // a_elkan.point_center_dist_2( x1, cndx)
+static PyObject * Drake_point_center_dist_2(DrakeObject *self, PyObject *args) {
+    // a_drake.point_center_dist_2( x1, cndx)
 
     int x1;
     unsigned short cndx;
-
     if (!PyArg_ParseTuple(args, "iH", &x1, &cndx)) {
         return NULL;
     }
@@ -262,11 +265,10 @@ static PyObject * Elkan_point_center_dist_2(ElkanObject *self, PyObject *args) {
     return PyFloat_FromDouble(dist2);
 }
 
-static PyObject * Elkan_center_center_dist_2(ElkanObject *self, PyObject *args) {
-    // a_elkan.center_center_dist_2(c1, c2)
+static PyObject * Drake_center_center_dist_2(DrakeObject *self, PyObject *args) {
+    // a_drake.center_center_dist_2(c1, c2)
 
     unsigned short c1, c2;
-
     if (!PyArg_ParseTuple(args, "HH", &c1, &c2)) {
         return NULL;
     }
@@ -277,70 +279,117 @@ static PyObject * Elkan_center_center_dist_2(ElkanObject *self, PyObject *args) 
 }
 
 
-// Elkan method definitions
+// Drake method definitions
 
 
-static PyMethodDef Elkan_methods[] = {
-    {"run", (PyCFunction) Elkan_run, METH_VARARGS | METH_KEYWORDS,
+static PyMethodDef Drake_methods[] = {
+    {"run", (PyCFunction) Drake_run, METH_VARARGS | METH_KEYWORDS,
         "Run threads until convergence or max iters, and returns num iters"},
-    {"free", (PyCFunction) Elkan_free, METH_NOARGS, "Free the object's memory"},
-    {"initialize", (PyCFunction) Elkan_initialize,
+    {"free", (PyCFunction) Drake_free, METH_NOARGS, "Free the object's memory"},
+    {"initialize", (PyCFunction) Drake_initialize,
         METH_VARARGS | METH_KEYWORDS,
         "Initialize algorithm at beginning of run() with given data and "
             "initial_assignment, which will be modified to contain final "
             "assignment of clusters"},
     {"point_point_inner_product",
-        (PyCFunction) Elkan_point_point_inner_product,
+        (PyCFunction) Drake_point_point_inner_product,
         METH_VARARGS,
         "Compute inner product. Could be standard dot operator, or kernel "
             "function for more exotic applications."},
     {"point_center_inner_product",
-        (PyCFunction) Elkan_point_center_inner_product,
+        (PyCFunction) Drake_point_center_inner_product,
         METH_VARARGS,
         "Compute inner product. Could be standard dot operator, or kernel "
             "function for more exotic applications."},
     {"center_center_inner_product",
-        (PyCFunction) Elkan_center_center_inner_product,
+        (PyCFunction) Drake_center_center_inner_product,
         METH_VARARGS,
         "Compute inner product. Could be standard dot operator, or kernel "
             "function for more exotic applications."},
-    {"point_center_dist_2", (PyCFunction) Elkan_point_center_dist_2,
+    {"point_center_dist_2", (PyCFunction) Drake_point_center_dist_2,
         METH_VARARGS,
         "Use the inner products to computer squared distances between a point "
             "and center."},
-    {"center_center_dist_2", (PyCFunction) Elkan_center_center_dist_2,
+    {"center_center_dist_2", (PyCFunction) Drake_center_center_dist_2,
         METH_VARARGS,
         "Use the inner products to computer squared distances between two "
             "centers."},
-    {"get_assignment", (PyCFunction) Elkan_get_assignment, METH_VARARGS,
+    {"get_assignment", (PyCFunction) Drake_get_assignment, METH_VARARGS,
         "Get the cluster assignment for the given point index"},
-    {"verify_assignment", (PyCFunction) Elkan_verify_assignment, METH_VARARGS,
+    {"verify_assignment", (PyCFunction) Drake_verify_assignment, METH_VARARGS,
         "Verify that current assignment is correct, by checking every "
             "point-center distance. For debugging."},
-    {"get_sse", (PyCFunction) Elkan_get_sse, METH_NOARGS,
+    {"get_sse", (PyCFunction) Drake_get_sse, METH_NOARGS,
         "Return the sum of squared errors for each cluster"},
-    {"get_name", (PyCFunction) Elkan_get_name, METH_NOARGS,
+    {"get_name", (PyCFunction) Drake_get_name, METH_NOARGS,
         "Return the algorithm name"},
     {NULL} // Sentinel
 };
 
 
-// Elkan type object
+// Drake type object
 
 
-PyTypeObject ElkanType = {
+PyTypeObject DrakeType = {
     PyVarObject_HEAD_INIT(NULL, 0)
-};
+    "fastkmeans.Drake", // tp_name
+    sizeof(DrakeObject), // tp_basicsize
+    0, // tp_itemsize
 
-void init_elkan_type_fields(void) {
-    ElkanType.tp_name = "fastkmeans.Elkan";
-    ElkanType.tp_doc = "";
-    ElkanType.tp_basicsize = sizeof(ElkanObject);
-    ElkanType.tp_itemsize = 0;
-    ElkanType.tp_flags = Py_TPFLAGS_DEFAULT | Py_TPFLAGS_BASETYPE;
-    ElkanType.tp_new = PyType_GenericNew;
-    ElkanType.tp_init = (initproc) Elkan_init;
-    ElkanType.tp_dealloc = (destructor) Elkan_dealloc;
-    ElkanType.tp_methods = Elkan_methods;
-    ElkanType.tp_getset = Elkan_getsetters;
+    (destructor) Drake_dealloc, // tp_dealloc
+    NULL, // tp_print
+    NULL, // tp_getattr
+    NULL, // tp_setattr
+    NULL, // tp_as_sync
+    NULL, // tp_repr
+
+    NULL, // tp_as_number
+    NULL, // tp_as_sequence
+    NULL, // tp_as_mapping
+
+    NULL, // tp_hash
+    NULL, // tp_call TODO ?
+    NULL, // tp_str
+    NULL, // tp_getattro
+    NULL, // tp_setattro
+
+    NULL, // tp_as_buffer
+
+    Py_TPFLAGS_DEFAULT | Py_TPFLAGS_BASETYPE, // tp_flags
+
+    "", // tp_doc
+
+    NULL, // tp_traverse
+
+    NULL, // tp_clear
+
+    NULL, // tp_richcompare
+
+    0, // tp_weaklistoffset
+
+    NULL, // tp_iter
+    NULL, // tp_iternext
+
+    Drake_methods, // tp_methods
+    NULL, // tp_members
+    Drake_getsetters, // tp_getset
+    NULL, // tp_base
+    NULL, // tp_dict
+    NULL, // tp_descr_get
+    NULL, // tp_descr_set
+    0, // tp_dictoffset
+    (initproc) Drake_init, // tp_init
+    PyType_GenericAlloc, // tp_alloc
+    PyType_GenericNew, // tp_new
+    NULL, // tp_free
+    NULL, // tp_is_gc
+    NULL, // tp_bases
+    NULL, // tp_mro
+    NULL, // tp_cache
+    NULL, // tp_subclasses
+    NULL, // tp_weaklist
+    NULL, // tp_del
+
+    0, // tp_version_tag
+    NULL, // tp_finalize
 };
