+++ src/python-bindings/py_fastkmeans_methods.h
@@ -1,0 +1,13 @@
+#ifndef PY_FASTKMEANS_METHODS_H
+#define PY_FASTKMEANS_METHODS_H
+
+#include <Python.h>
+#include <structmember.h>
+
+/* Provides a wrapper for the general functions. See general_functions.h for
+ * more detail.
+ */
+
+extern PyMethodDef Fastkmeans_methods[];
+
+#endif
