+++ src/python-bindings/py_heap.cpp
@@ -1,17 +1,20 @@
-#include "py_annulus.h"
-#include "py_dataset.h"
+/* HeapKmeans wrapper. The comment at the beginning of each function definition
+ * demonstrates its usage in Python.
+ */
+
+#include "py_heap.h"
+
 #include "py_assignment.h"
+#include "py_dataset.h"
 
-//#include "annulus_kmeans.h"
 
-
-// Annulus instance object
+// Heap instance object
 
 
 /*
 typedef struct {
     PyObject_HEAD
-    AnnulusKmeans *instance;
+    HeapKmeans *instance;
 
     PyObject *dataset;
     PyObject *assignment;
@@ -20,18 +23,17 @@ typedef struct {
     // #ifdef COUNT_DISTANCES
     // long long num_distances;
     // #endif
-} AnnulusObject;
+} HeapObject;
 */
 
 
 // Object special methods
 
 
-static int Annulus_init(AnnulusObject *self) {
-    // Annulus()
+static int Heap_init(HeapObject *self) {
+    // Heap()
 
-    self->instance = new AnnulusKmeans();
-
+    self->instance = new HeapKmeans();
     Py_INCREF(Py_None);
     self->dataset = Py_None;
     Py_INCREF(Py_None);
@@ -40,9 +42,8 @@ static int Annulus_init(AnnulusObject *self) {
     return 0;
 }
 
-static void Annulus_dealloc(AnnulusObject *self) {
+static void Heap_dealloc(HeapObject *self) {
     delete self->instance;
-
     Py_DECREF(self->dataset);
     self->dataset = NULL;
     Py_DECREF(self->assignment);
@@ -55,8 +56,8 @@ static void Annulus_dealloc(AnnulusObject *self) {
 // Object properties
 
 
-static PyObject * Annulus_get_centers(AnnulusObject *self, void *closure) {
-    // a_annulus.centers
+static PyObject * Heap_get_centers(HeapObject *self, void *closure) {
+    // a_heap.centers
 
     const Dataset *centers = self->instance->getCenters();
     if (centers == NULL) {
@@ -73,43 +74,45 @@ static PyObject * Annulus_get_centers(AnnulusObject *self, void *closure) {
 
     // Copy values from centers to preserve constness
 
-    // Dataset *newCenters = new Dataset(centers->n, centers->d);
     for (int i = 0; i < centers->nd; i++) {
-        //newCenters->data[i] = centers->data[i];
         centersObj->dataset->data[i] = centers->data[i];
     }
-
-    //centersObj->dataset = newCenters;
 
     return (PyObject *) centersObj;
 }
 
-static PyGetSetDef Annulus_getsetters[] = {
-    // Readonly
-    {"centers", (getter) Annulus_get_centers, NULL, "The set of centers"},
+static PyGetSetDef Heap_getsetters[] = {
+    {
+        const_cast<char *>("centers"),
+        (getter) Heap_get_centers,
+        NULL, // setter
+        const_cast<char *>("The set of centers")
+    },
     {NULL} // Sentinel
 };
 
 
-// AnnulusKmeans methods
+// HeapKmeans methods
 
 
-static PyObject * Annulus_free(AnnulusObject *self) {
-    // a_annulus.free()
+static PyObject * Heap_free(HeapObject *self) {
+    // a_heap.free()
 
     self->instance->free();
     Py_RETURN_NONE;
 }
 
-static PyObject * Annulus_initialize(AnnulusObject *self, PyObject *args,
+static PyObject * Heap_initialize(HeapObject *self, PyObject *args,
         PyObject *kwargs) {
-    // a_annulus.initialize(x, k, initial_assignment, num_threads=an_int)
+    // a_heap.initialize(x, k, initial_assignment, num_threads=an_int)
 
     PyObject *x_orig, *initAssigns_orig;
     unsigned short k;
     int numThreads = 1;
 
-    char *kwlist[] = {"", "", "", "num_threads", NULL};
+    char *emptyStr = const_cast<char *>("");
+    char *kwlist[] = {emptyStr, emptyStr, emptyStr,
+        const_cast<char *>("num_threads"), NULL};
 
     if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O!HO!|i", kwlist,
                 &DatasetType, &x_orig, &k, &AssignmentType, &initAssigns_orig,
@@ -117,10 +120,12 @@ static PyObject * Annulus_initialize(AnnulusObject *self, PyObject *args,
         return NULL;
     }
 
+    // Reassign dataset
     Py_DECREF(self->dataset);
     Py_INCREF(x_orig);
     self->dataset = x_orig;
 
+    // Reassign assignment
     Py_DECREF(self->assignment);
     Py_INCREF(initAssigns_orig);
     self->assignment = initAssigns_orig;
@@ -137,8 +142,8 @@ static PyObject * Annulus_initialize(AnnulusObject *self, PyObject *args,
     Py_RETURN_NONE;
 }
 
-static PyObject * Annulus_get_name(AnnulusObject *self) {
-    // a_annulus.get_name()
+static PyObject * Heap_get_name(HeapObject *self) {
+    // a_heap.get_name()
 
     return PyUnicode_FromString(self->instance->getName().c_str());
 }
@@ -147,12 +152,11 @@ static PyObject * Annulus_get_name(AnnulusObject *self) {
 // OriginalSpaceKmeans methods
 
 
-static PyObject * Annulus_point_point_inner_product(AnnulusObject *self,
+static PyObject * Heap_point_point_inner_product(HeapObject *self,
         PyObject *args) {
-    // a_annulus.point_point_inner_product(x1, x2)
+    // a_heap.point_point_inner_product(x1, x2)
 
     int x1ndx, x2ndx;
-
     if (!PyArg_ParseTuple(args, "ii", &x1ndx, &x2ndx)) {
         return NULL;
     }
@@ -162,13 +166,12 @@ static PyObject * Annulus_point_point_inner_product(AnnulusObject *self,
     return PyFloat_FromDouble(innerProd);
 }
 
-static PyObject * Annulus_point_center_inner_product(AnnulusObject *self,
+static PyObject * Heap_point_center_inner_product(HeapObject *self,
         PyObject *args) {
-    // a_annulus.point_center_inner_product(xndx, cndx)
+    // a_heap.point_center_inner_product(xndx, cndx)
 
     int xndx;
     unsigned short cndx;
-
     if (!PyArg_ParseTuple(args, "iH", &xndx, &cndx)) {
         return NULL;
     }
@@ -178,12 +181,11 @@ static PyObject * Annulus_point_center_inner_product(AnnulusObject *self,
     return PyFloat_FromDouble(innerProd);
 }
 
-static PyObject * Annulus_center_center_inner_product(AnnulusObject *self,
+static PyObject * Heap_center_center_inner_product(HeapObject *self,
         PyObject *args) {
-    // a_annulus.center_center_inner_product(c1, c2)
+    // a_heap.center_center_inner_product(c1, c2)
 
     unsigned short c1, c2;
-
     if (!PyArg_ParseTuple(args, "HH", &c1, &c2)) {
         return NULL;
     }
@@ -197,13 +199,12 @@ static PyObject * Annulus_center_center_inner_product(AnnulusObject *self,
 // Kmeans methods
 
 
-static PyObject * Annulus_run(AnnulusObject *self, PyObject *args,
+static PyObject * Heap_run(HeapObject *self, PyObject *args,
         PyObject *kwargs) {
-    // a_annulus.run(max_iterations = 0)
+    // a_heap.run(max_iterations = 0)
 
-    static char *kwlist[] = {"max_iterations"};
     int maxIterations = 0;
-
+    static char *kwlist[] = {const_cast<char *>("max_iterations"), NULL};
     if (!PyArg_ParseTupleAndKeywords(args, kwargs, "|i", kwlist,
                 &maxIterations)) {
         return NULL;
@@ -215,11 +216,10 @@ static PyObject * Annulus_run(AnnulusObject *self, PyObject *args,
     return PyLong_FromLong(numIterations);
 }
 
-static PyObject * Annulus_get_assignment(AnnulusObject *self, PyObject *args) {
-    // a_annulus.get_assignment(xndx)
+static PyObject * Heap_get_assignment(HeapObject *self, PyObject *args) {
+    // a_heap.get_assignment(xndx)
 
     int xndx;
-
     if (!PyArg_ParseTuple(args, "i", &xndx)) {
         return NULL;
     }
@@ -229,11 +229,10 @@ static PyObject * Annulus_get_assignment(AnnulusObject *self, PyObject *args) {
     return PyLong_FromLong(assignment);
 }
 
-static PyObject * Annulus_verify_assignment(AnnulusObject *self, PyObject *args) {
-    // a_annulus.verify_assignment(iteration, startndx, endndx)
+static PyObject * Heap_verify_assignment(HeapObject *self, PyObject *args) {
+    // a_heap.verify_assignment(iteration, startndx, endndx)
 
     int iteration, startNdx, endNdx;
-
     if (!PyArg_ParseTuple(args, "iii", &iteration, &startNdx, &endNdx)) {
         return NULL;
     }
@@ -241,18 +240,17 @@ static PyObject * Annulus_verify_assignment(AnnulusObject *self, PyObject *args)
     Py_RETURN_NONE;
 }
 
-static PyObject * Annulus_get_sse(AnnulusObject *self) {
-    // a_annulus.get_sse()
+static PyObject * Heap_get_sse(HeapObject *self) {
+    // a_heap.get_sse()
 
     return PyFloat_FromDouble(self->instance->getSSE());
 }
 
-static PyObject * Annulus_point_center_dist_2(AnnulusObject *self, PyObject *args) {
-    // a_annulus.point_center_dist_2( x1, cndx)
+static PyObject * Heap_point_center_dist_2(HeapObject *self, PyObject *args) {
+    // a_heap.point_center_dist_2( x1, cndx)
 
     int x1;
     unsigned short cndx;
-
     if (!PyArg_ParseTuple(args, "iH", &x1, &cndx)) {
         return NULL;
     }
@@ -262,11 +260,10 @@ static PyObject * Annulus_point_center_dist_2(AnnulusObject *self, PyObject *arg
     return PyFloat_FromDouble(dist2);
 }
 
-static PyObject * Annulus_center_center_dist_2(AnnulusObject *self, PyObject *args) {
-    // a_annulus.center_center_dist_2(c1, c2)
+static PyObject * Heap_center_center_dist_2(HeapObject *self, PyObject *args) {
+    // a_heap.center_center_dist_2(c1, c2)
 
     unsigned short c1, c2;
-
     if (!PyArg_ParseTuple(args, "HH", &c1, &c2)) {
         return NULL;
     }
@@ -277,70 +274,117 @@ static PyObject * Annulus_center_center_dist_2(AnnulusObject *self, PyObject *ar
 }
 
 
-// Annulus method definitions
+// Heap method definitions
 
 
-static PyMethodDef Annulus_methods[] = {
-    {"run", (PyCFunction) Annulus_run, METH_VARARGS | METH_KEYWORDS,
+static PyMethodDef Heap_methods[] = {
+    {"run", (PyCFunction) Heap_run, METH_VARARGS | METH_KEYWORDS,
         "Run threads until convergence or max iters, and returns num iters"},
-    {"free", (PyCFunction) Annulus_free, METH_NOARGS, "Free the object's memory"},
-    {"initialize", (PyCFunction) Annulus_initialize,
+    {"free", (PyCFunction) Heap_free, METH_NOARGS, "Free the object's memory"},
+    {"initialize", (PyCFunction) Heap_initialize,
         METH_VARARGS | METH_KEYWORDS,
         "Initialize algorithm at beginning of run() with given data and "
             "initial_assignment, which will be modified to contain final "
             "assignment of clusters"},
     {"point_point_inner_product",
-        (PyCFunction) Annulus_point_point_inner_product,
+        (PyCFunction) Heap_point_point_inner_product,
         METH_VARARGS,
         "Compute inner product. Could be standard dot operator, or kernel "
             "function for more exotic applications."},
     {"point_center_inner_product",
-        (PyCFunction) Annulus_point_center_inner_product,
+        (PyCFunction) Heap_point_center_inner_product,
         METH_VARARGS,
         "Compute inner product. Could be standard dot operator, or kernel "
             "function for more exotic applications."},
     {"center_center_inner_product",
-        (PyCFunction) Annulus_center_center_inner_product,
+        (PyCFunction) Heap_center_center_inner_product,
         METH_VARARGS,
         "Compute inner product. Could be standard dot operator, or kernel "
             "function for more exotic applications."},
-    {"point_center_dist_2", (PyCFunction) Annulus_point_center_dist_2,
+    {"point_center_dist_2", (PyCFunction) Heap_point_center_dist_2,
         METH_VARARGS,
         "Use the inner products to computer squared distances between a point "
             "and center."},
-    {"center_center_dist_2", (PyCFunction) Annulus_center_center_dist_2,
+    {"center_center_dist_2", (PyCFunction) Heap_center_center_dist_2,
         METH_VARARGS,
         "Use the inner products to computer squared distances between two "
             "centers."},
-    {"get_assignment", (PyCFunction) Annulus_get_assignment, METH_VARARGS,
+    {"get_assignment", (PyCFunction) Heap_get_assignment, METH_VARARGS,
         "Get the cluster assignment for the given point index"},
-    {"verify_assignment", (PyCFunction) Annulus_verify_assignment, METH_VARARGS,
+    {"verify_assignment", (PyCFunction) Heap_verify_assignment, METH_VARARGS,
         "Verify that current assignment is correct, by checking every "
             "point-center distance. For debugging."},
-    {"get_sse", (PyCFunction) Annulus_get_sse, METH_NOARGS,
+    {"get_sse", (PyCFunction) Heap_get_sse, METH_NOARGS,
         "Return the sum of squared errors for each cluster"},
-    {"get_name", (PyCFunction) Annulus_get_name, METH_NOARGS,
+    {"get_name", (PyCFunction) Heap_get_name, METH_NOARGS,
         "Return the algorithm name"},
     {NULL} // Sentinel
 };
 
 
-// Annulus type object
+// Heap type object
 
 
-PyTypeObject AnnulusType = {
+PyTypeObject HeapType = {
     PyVarObject_HEAD_INIT(NULL, 0)
-};
+    "fastkmeans.Heap", // tp_name
+    sizeof(HeapObject), // tp_basicsize
+    0, // tp_itemsize
 
-void init_annulus_type_fields(void) {
-    AnnulusType.tp_name = "fastkmeans.Annulus";
-    AnnulusType.tp_doc = "";
-    AnnulusType.tp_basicsize = sizeof(AnnulusObject);
-    AnnulusType.tp_itemsize = 0;
-    AnnulusType.tp_flags = Py_TPFLAGS_DEFAULT | Py_TPFLAGS_BASETYPE;
-    AnnulusType.tp_new = PyType_GenericNew;
-    AnnulusType.tp_init = (initproc) Annulus_init;
-    AnnulusType.tp_dealloc = (destructor) Annulus_dealloc;
-    AnnulusType.tp_methods = Annulus_methods;
-    AnnulusType.tp_getset = Annulus_getsetters;
+    (destructor) Heap_dealloc, // tp_dealloc
+    NULL, // tp_print
+    NULL, // tp_getattr
+    NULL, // tp_setattr
+    NULL, // tp_as_sync
+    NULL, // tp_repr
+
+    NULL, // tp_as_number
+    NULL, // tp_as_sequence
+    NULL, // tp_as_mapping
+
+    NULL, // tp_hash
+    NULL, // tp_call TODO ?
+    NULL, // tp_str
+    NULL, // tp_getattro
+    NULL, // tp_setattro
+
+    NULL, // tp_as_buffer
+
+    Py_TPFLAGS_DEFAULT | Py_TPFLAGS_BASETYPE, // tp_flags
+
+    "", // tp_doc
+
+    NULL, // tp_traverse
+
+    NULL, // tp_clear
+
+    NULL, // tp_richcompare
+
+    0, // tp_weaklistoffset
+
+    NULL, // tp_iter
+    NULL, // tp_iternext
+
+    Heap_methods, // tp_methods
+    NULL, // tp_members
+    Heap_getsetters, // tp_getset
+    NULL, // tp_base
+    NULL, // tp_dict
+    NULL, // tp_descr_get
+    NULL, // tp_descr_set
+    0, // tp_dictoffset
+    (initproc) Heap_init, // tp_init
+    PyType_GenericAlloc, // tp_alloc
+    PyType_GenericNew, // tp_new
+    NULL, // tp_free
+    NULL, // tp_is_gc
+    NULL, // tp_bases
+    NULL, // tp_mro
+    NULL, // tp_cache
+    NULL, // tp_subclasses
+    NULL, // tp_weaklist
+    NULL, // tp_del
+
+    0, // tp_version_tag
+    NULL, // tp_finalize
 };
