+++ src/python-bindings/py_sort.h
@@ -1,15 +1,18 @@
-#ifndef PY_NAIVE_H
-#define PY_NAIVE_H
+#ifndef PY_SORT_H
+#define PY_SORT_H
+
+/* Provides a wrapper for the SortKmeans class. See sort_kmeans.h for more
+ * detail.
+ */
 
 #include <Python.h>
 #include <structmember.h>
 
-#include "naive_kmeans.h"
+#include "sort_kmeans.h"
 
 typedef struct {
     PyObject_HEAD
-    NaiveKmeans *instance;
-    // TODO make these visible, probably readonly
+    SortKmeans *instance;
     PyObject *dataset;
     PyObject *assignment;
 
@@ -17,10 +20,8 @@ typedef struct {
     // #ifdef COUNT_DISTANCES
     // long long num_distances;
     // #endif
-} NaiveObject;
+} SortObject;
 
-extern PyTypeObject NaiveType;
-
-void init_naive_type_fields(void);
+extern PyTypeObject SortType;
 
 #endif
